BtreeMap iterator should not return key/value tuples like suggested in the documentation.
The exercise asked for the ticket -- deduced since t.id requests an id field which
TicketId does not have.

To do this, we need to use std::collections:btree_map::Values<'a, TicketId, Ticket>
as the intoiter type. This makes it so that we can return the values of the BtreeMap
using self.tickets.values()